Changed the horizontal speed to be one half of the vertial speed. FIxed collision on the paddle so the ball will not clip inside of the top of it.

+++ src/ball.h
@@ -8,7 +8,7 @@
 class Ball {
 public:
 	Stats playerStats{0, 0, 99, false, false, false};
-	Ball(const Coordinate2D coord, const int radius, const int speed);
+	Ball(const Coordinate2D coord, const int radius, const float speed);
 	void draw();
 	void move(std::vector<std::vector<Brick>>& bricks, Paddle& paddle);
 	void checkWallCollision(Paddle& paddle);
@@ -18,5 +18,5 @@ private:
 	Coordinate2D coord;
 	Coordinate2D direction{ 1, 1 };
 	int radius;
-	int speed;
+	float speed;
 };
